+++ common/params.cc
@@ -320,6 +320,7 @@ std::unordered_map<std::string, uint32_t> keys = {
     { "TFollowGap2", PERSISTENT },
     { "TFollowGap3", PERSISTENT },
     { "TFollowGap4", PERSISTENT },
+    { "DynamicJerk", PERSISTENT },
     { "JerkStartLimit", PERSISTENT },
     { "JerkUpperLowerLimit", PERSISTENT },
     { "KeepEngage", PERSISTENT },